Add xmmap() and xmunmap() and use them instead of mmap() and munmap()
mmap() doesn't allow 0 size mapping which causes bugs when reading empty
files.  Return magic pointer which points to outside of the process
address space for 0 sized mappings.  xmunmap() does nothing if that
magic pointer is passed to it.

Return NULL if mmap() fails instead of MAP_FAILED which is more logical.