Increase maximum block size to 512 bytes
The maximum block size was only 64 bytes because it increased the
possibility of hitting bugs in low level code.  The code seems quite
stable now, so we can increase it to 512 bytes.

Maybe we should use large blocks by default and then split blocks into
smaller ones when making changes, but it could be quite tricky and
probably not worth the pain.