Move cursor to bottom of screen before switching back to main buffer
We must not move cursor after switching back to main buffer because it
messes up the screen.  If terminal supports separate buffer for
interactive programs (has termcap entries ti and te set) then we don't
need to move cursor at all, but because GNU Screen claims it supports
the feature but fails to emulate it we need to move cursor anyway.  By
doing that before switching back to the main buffer makes everyone
happy, hopefully.