Handle truncated escape sequences correctly
Try to guess if escape sequence has been truncated and wait for more
input.  This happens for example when scrolling through a big file in
gnome-terminal. Many keypresses are buffered and the last one gets
truncated.

Also wait one second for more input if there's only ESC in the input
buffer.  The delay could be shorter but one second is enough for you
to type another character so this make it possible to type ESC x to
emulate M-x.