Add builtin ANSI escape codes for arrow keys
Normally when keypad is on we only need to support the escape codes
defined in terminfo entry.  But when keypad is off terminal sends ANSI
escape codes for arrow keys instead.  This happens at least when using
dtach to attach to existing session without using "-r winch" option
(editor sends "keypad on" initialization code to terminal after window
resize but not when it receives ^L from dtach).