Configurable whitespace error highlighting
Remove indentation error and trailing whitespace highlighting from
syntax files and add special configurable highlighter for this purpose.

This is much more flexible because you can specify which whitespace
errors are highlighter for each buffer by changing the new ws-error
option.  The option is global and local (global value used as a default
for new buffers) and is a bit-mask consisting of a sum of any values
below.  Set to 0 to disable completely.

Value  Highlighted as an error
------------------------------------------------------------------------
    1  Trailing whitespace.  Don't allow tabs or spaces at end of a
       line.

    2  Spaces in indentation.  Does not include less than tab-width
       spaces at end of indentation which are commonly used to align
       function parameters in C.

    4  Less than tab-width spaces at end of indentation.  Don't allow
       any spaces at end of indentation.  Used sometimes with the value
       above (2), quite useless alone.

    8  Tab in indentation.  Only spaces allowed for indentation.

   16  Tab anywhere but in indentation.  Could be useful detect
       (incorrect) use of tabs in anywhere but indentation when
       tab-width is set to any other value than 8.