Support environment variables in commands
In addition following special variables are recognized:

    DATADIR    The variable set at build time.
    FILENAME   Absolute name of currently edited file.
    WORD       Word under cursor.

Variables are expanded when executed and can't be inside quotes.  Unlike
in shell, a variable can contain whitespace and it will still be treated
as a single parameter.

Example:
    alias x "run chmod 755 $FILENAME"

When "x" is executed $FILENAME is expanded to current filename.

Incorrect usage example:
    alias z 'run -p sh -c "echo $FILENAME"'

In the example above $FILENAME would not be expanded because it is
inside quotes.  So the sh command would see $FILENAME and expand it
(most likely) to empty string.