Use int instead of pid_t
I was wondering what type pit_t really is so I looked at the GNU libc
header files and puked.

/usr/include/unistd.h
	typedef __pid_t pid_t;

/usr/include/bits/types.h
	# define __STD_TYPE             typedef
	__STD_TYPE __PID_T_TYPE __pid_t;        /* Type of process identifications.  */

/usr/include/bits/typesizes.h
	#define __PID_T_TYPE            __S32_TYPE

/usr/include/bits/types.h
	#define __S32_TYPE              int

So it appears that pid_t is always int on systems using glibc.  If it's
long on some other system you'll get a compiler warning but it will
still work and is more readable.