Syntax file loading changes
Separate loading by filename and by file type.

Change load-syntax to take only one parameter. If the parameter contains
slash then it is treated as a filename, otherwise it is file type.

When loading by file type it is not an error if the syntax doesn't exist
or has already been loaded.

When loading by filename it is an error if the file does not exist or
syntax for the file type has already been loaded.