Add block_iter_eat_line()
It is similar to block_iter_next_line() but there are some differences:

    1) The iterator is moved after first seen newline even if it's the
       last byte of the file. In that case iterator is placed on the so
       called "dummy line".

    2) If there is no newline, then the iterator will be moved to end of
       the file.

    3) Return value is always number of bytes the iterator advanced.

The intention is to convert some block_iter_next_line() callers to use
this new function and then change block_iter_next_line() to never change
offset if return value is 0. This behavior would be much easier to
understand.