Consistent file open behavior
Make git-open behave more like the open command.

Close empty/untouched buffer if it was the only buffer in the window.
Remember previous view.

Also change "open" to not remember previous view when opening multiple
files. That means that "open str.c str.h; close" activates str.h, not
the view that was active before the command was run.